requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.27
sqlmodel>=0.0.16
passlib[bcrypt]>=1.7
bcrypt>=4.0.1
PyJWT>=2.8.0
